requirements.txt
altair==6.0.0
appscript==1.3.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.4.0
babel==2.17.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.3
cryptography==44.0.2
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
Django==5.1.2
django-axes==8.3.1
django-cors-headers==4.7.0
django-environ==0.13.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
docx2pdf==0.1.8
et_xmlfile==2.0.0
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
html5lib==1.1
httplib2==0.22.0
idna==3.10
inflect==7.5.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lml==0.2.0
lxml==5.3.0
MarkupSafe==3.0.3
matplotlib==3.10.8
more-itertools==10.8.0
mysqlclient==2.2.5
narwhals==2.14.0
nose==1.3.7
num2words==0.5.14
numpy==2.4.0
oauthlib==3.2.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pillow==11.0.0
plotly==6.5.0
proto-plus==1.26.1
protobuf==6.30.2
py-dateutil==2.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydeck==0.9.1
pyexcel==0.7.3
pyexcel-io==0.6.7
pyexcel-xlsx==0.6.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
pypdf==5.4.0
pyproj==3.7.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
PyYAML==6.0.2
qrcode==8.1
referencing==0.37.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9
selenium==4.27.1
six==1.17.0
smmap==5.0.2
sqlparse==0.5.1
streamlit==1.52.2
svglib==1.5.1
tenacity==9.1.2
texttable==1.7.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.4
tqdm==4.66.6
typeguard==4.4.4
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
whitenoise==6.10.0
xhtml2pdf==0.2.17
xlsxwriter==3.2.9
